pyproject.toml
[tool.poetry]

name = "ballotshare"
version = "0.0"
description = "Share how you plan to vote in Michigan."
authors = ["Citizen Labs <hello@citizenlabs.org>"]
license = "MIT"

packages = [
    { include = "app" },
]

[tool.poetry.dependencies]

python = "^3.7"

# Quart
quart = "~0.10.0"

# Images
pillow = "^7.0"

# Utilities
aiohttp = "^3.6.2"
minilog = "*"

[tool.poetry.dev-dependencies]

# Formatting
isort = "=4.3.21"
black = "=19.10b0"

# Type Checking
mypy = "*"
pylint = "*"

# Testing
pytest = "^5.0"
pytest-asyncio = "*"
pytest-describe = "^0.12.0"
pytest-expecter = "^1.3"
pytest-cov = "^2.7"
pytest-watch = "^4.2"

# Coverage
coveragespace = "^3.0.1"

# Utilities
honcho = "^1.0"
rope = "^0.14.0"

[build-system]

requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
